fix: address Claude Code review round 18 findings (#18)
* fix: address Claude Code review round 18 - 4 actionable fixes

L-1: Move regexp.MustCompile to package-level var (avoid per-call recompile)
M-1: Change writeMu.RLock() to writeMu.Lock() - serialize write handlers to prevent read-modify-write race on same card
M-2: Add tag count limit (max 100) in validateNote
L-5: Fix protobuf uint64->int casts on 32-bit overflow (use uint64 comparison)

Not-fixed (by-design or already clean):
L-2 (SQL keyword check) - already fixed in PR #17
L-3 (HTML unescaped) - by design, Anki cards contain HTML
L-4 (rows not deferred) - manual close is correct on all paths, panics handled by recoverPanic
L-6 (both .anki2 and .anki21b) - by design, .anki21b wins

* fix: update stale comments after writeMu lock upgrade

+++ pkg/collection/schema.go
@@ -255,7 +255,7 @@ func parseTemplateConfig(data []byte) (qfmt, afmt string) {
 			data = data[8:]
 		case 2: // length-delimited
 			length, n := decodeVarint(data)
-			if n <= 0 || int(length) > len(data[n:]) || int(length) > maxProtobufFieldLen {
+			if n <= 0 || length > uint64(len(data[n:])) || length > uint64(maxProtobufFieldLen) {
 				break
 			}
 			data = data[n:]
@@ -321,7 +321,7 @@ func parseNotetypeConfig(data []byte, m goanki.Model) goanki.Model {
 			data = data[8:]
 		case 2: // length-delimited
 			length, n := decodeVarint(data)
-			if n <= 0 || int(length) > len(data[n:]) || int(length) > maxProtobufFieldLen {
+			if n <= 0 || length > uint64(len(data[n:])) || length > uint64(maxProtobufFieldLen) {
 				break
 			}
 			data = data[n:]
